requirements.txt
numpy
scipy
pandas
matplotlib
opencv-python
scikit-learn
torch==1.8.0
torchvision==0.9.0
efficientnet_pytorch
tqdm
flask
Pillow
gunicorn